requirements.txt
gradio
langchain
langchain-groq
python-dotenv
openpyxl
pandas
faiss-cpu
sentence-transformers
langchain-community